+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to
 - `owner` and `is_main` fields on `CreditCard` model are deprecated
   and been replaced with many-to-many `owners` field instead
 
+### Changed
+
+- Delete credit card once all orders installments are paid
+
 ## [2.15.0] - 2025-02-05
 
 ### Changed